Executors - Services - Console ON WS

+++ API/src/main/java/be/alexandre01/dreamnetwork/api/utils/messages/mapper/MapperOfPlayer.java
@@ -11,15 +11,15 @@ import java.util.Date;
  ↬   Made by Alexandre01Dev 😎
  ↬   done on 05/11/2023 at 20:02
 */
-public class MapperOfPlayer extends ObjectConverterMapper<Player, Integer> {
+public class MapperOfPlayer extends ObjectConverterMapper<Player, Long> {
 
     @Override
-    public Integer convert(Player player) {
+    public Long convert(Player player) {
         return player.getId();
     }
 
     @Override
-    public Player read(Integer id) {
+    public Player read(Long id) {
         return DNCoreAPI.getInstance().getServicePlayersManager().getPlayer(id);
     }
 }
